+++ Tests/ProjectDescriptionTests/PresetBuildConfigurationTests.swift
@@ -1,0 +1,12 @@
+import Foundation
+import TuistCoreTesting
+import XCTest
+
+@testable import ProjectDescription
+
+final class PresetBuildConfigurationTests: XCTestCase {
+    func test_preset_names() {
+        XCTAssertEqual(PresetBuildConfiguration.debug.name, "Debug")
+        XCTAssertEqual(PresetBuildConfiguration.release.name, "Release")
+    }
+}