+++ Tests/ProjectDescriptionTests/TargetTests.swift
@@ -70,18 +70,28 @@ final class TargetTests: XCTestCase {
                 "base": {
                     "a": "b"
                 },
-                "debug": {
-                    "xcconfig": "config",
-                    "settings": {
-                        "a": "b"
-                    }
-                },
-                "release": {
-                    "xcconfig": "config",
-                    "settings": {
-                        "a": "b"
-                    }
-                },
+                "configurations": [
+                    {
+                        "name": "Debug",
+                        "variant": "debug",
+                        "configuration": {
+                            "xcconfig": "config",
+                            "settings": {
+                                "a": "b"
+                            }
+                        }
+                    },
+                    {
+                        "name": "Release",
+                        "variant": "release",
+                        "configuration": {
+                            "xcconfig": "config",
+                            "settings": {
+                                "a": "b"
+                            }
+                        }
+                    },
+                ],
                 "defaultSettings": "recommended"
             },
             "resources": [
@@ -191,18 +201,28 @@ final class TargetTests: XCTestCase {
                 "base": {
                     "a": "b"
                 },
-                "debug": {
-                    "xcconfig": "config",
-                    "settings": {
-                        "a": "b"
-                    }
-                },
-                "release": {
-                    "xcconfig": "config",
-                    "settings": {
-                        "a": "b"
-                    }
-                },
+                "configurations": [
+                    {
+                        "name": "Debug",
+                        "variant": "debug",
+                        "configuration": {
+                            "xcconfig": "config",
+                            "settings": {
+                                "a": "b"
+                            }
+                        }
+                    },
+                    {
+                        "name": "Release",
+                        "variant": "release",
+                        "configuration": {
+                            "xcconfig": "config",
+                            "settings": {
+                                "a": "b"
+                            }
+                        }
+                    },
+                ],
                 "defaultSettings": "recommended"
             },
             "resources": [