+++ features/step_definitions/shared/xcode.rb
@@ -22,3 +22,18 @@ Then(/I should be able to (.+) the scheme (.+)/) do |action, scheme|
 
   xcodebuild(*args)
 end
+
+Then(/the scheme (.+) has a build setting (.+) with value (.+) for the configuration (.+)/) do |scheme, key, value, config|
+  args = [
+    "-scheme", scheme,
+    "-workspace", @workspace_path,
+    "-configuration", config,
+    "-showBuildSettings"
+  ]
+
+  out, err, status = Open3.capture3("xcodebuild", *args)
+  flunk(err) unless status.success?
+
+  search_for = "#{key} = #{value}"
+  assert(out.include?(search_for), "Couldn't find '#{search_for}'")
+end